requirements.txt
configparser
markdown
pafy
pydub

requests
tornado
pycurl
pyyaml
youtube-dl
